Clean up log implementation.
